preprocessing int konverzió pótlása

+++ src/preprocessing.py
@@ -36,7 +36,7 @@ def process_image(path):
     if shape[0] > dest_shape[0] and shape[1] > dest_shape[1]:
         scale = shape[0] / dest_shape[0]
         image = cv2.resize(image, dest_shape)
-        label_data_int = [item * scale for item in label_data_int]
+        label_data_int = [int(item * scale) for item in label_data_int]
     cv2.imwrite(f'{resdir}{"_".join(str(item) for item in label_data_int)}.png', image)
     return image
